pyscripts/requirements.txt
numpy
pandas
pyyaml
matplotlib
regex
lxml
scipy
bidict
seaborn